// src/model/proto/credential/account.credential.grpc.ts
export interface IAccountCredential {
  /** Document ID (_id in MongoDB) */
  id?:
    | string
    | undefined;
  /** Preference ObjectId (as string) */
  preference?:
    | string
    | undefined;
  /** Parent ObjectId (as string) */
  parent?:
    | IAccountCredentialParent
    | undefined;
  /** First name as string */
  email: string;
  /** Last name as string */
  username: string;
  password?: string | undefined;
}

/** The response message containing an AccountCredential */
export interface AccountCredentialReadRequest {
}

export interface AccountCredentialAuthRequest {
  username: string;
  password: string;
}

export interface AccountCredentialAuthResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential | undefined;
}

export interface AccountCredentialCreateResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential | undefined;
}

export interface AccountCredentialReadResponse {
  status: boolean;
  code: number;
  msg: string;
  data: IAccountCredential[];
}
// ...
export const AccountCredentialReadRequest: MessageFns<AccountCredentialReadRequest> = {
  encode(_: AccountCredentialReadRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialReadRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialReadRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(_: any): AccountCredentialReadRequest {
    return {};
  },

  toJSON(_: AccountCredentialReadRequest): unknown {
    const obj: any = {};
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialReadRequest>, I>>(base?: I): AccountCredentialReadRequest {
    return AccountCredentialReadRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialReadRequest>, I>>(_: I): AccountCredentialReadRequest {
    const message = createBaseAccountCredentialReadRequest();
    return message;
  },
};

function createBaseAccountCredentialAuthRequest(): AccountCredentialAuthRequest {
  return { username: "", password: "" };
}

export const AccountCredentialAuthRequest: MessageFns<AccountCredentialAuthRequest> = {
  encode(message: AccountCredentialAuthRequest, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.username !== "") {
      writer.uint32(10).string(message.username);
    }
    if (message.password !== "") {
      writer.uint32(18).string(message.password);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialAuthRequest {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialAuthRequest();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.username = reader.string();
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.password = reader.string();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountCredentialAuthRequest {
    return {
      username: isSet(object.username) ? globalThis.String(object.username) : "",
      password: isSet(object.password) ? globalThis.String(object.password) : "",
    };
  },

  toJSON(message: AccountCredentialAuthRequest): unknown {
    const obj: any = {};
    if (message.username !== "") {
      obj.username = message.username;
    }
    if (message.password !== "") {
      obj.password = message.password;
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialAuthRequest>, I>>(base?: I): AccountCredentialAuthRequest {
    return AccountCredentialAuthRequest.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialAuthRequest>, I>>(object: I): AccountCredentialAuthRequest {
    const message = createBaseAccountCredentialAuthRequest();
    message.username = object.username ?? "";
    message.password = object.password ?? "";
    return message;
  },
};

function createBaseAccountCredentialAuthResponse(): AccountCredentialAuthResponse {
  return { status: false, code: 0, msg: "", data: undefined };
}

export const AccountCredentialAuthResponse: MessageFns<AccountCredentialAuthResponse> = {
  encode(message: AccountCredentialAuthResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== false) {
      writer.uint32(8).bool(message.status);
    }
    if (message.code !== 0) {
      writer.uint32(16).int64(message.code);
    }
    if (message.msg !== "") {
      writer.uint32(26).string(message.msg);
    }
    if (message.data !== undefined) {
      IAccountCredential.encode(message.data, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialAuthResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialAuthResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.status = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.code = longToNumber(reader.int64());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.msg = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data = IAccountCredential.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountCredentialAuthResponse {
    return {
      status: isSet(object.status) ? globalThis.Boolean(object.status) : false,
      code: isSet(object.code) ? globalThis.Number(object.code) : 0,
      msg: isSet(object.msg) ? globalThis.String(object.msg) : "",
      data: isSet(object.data) ? IAccountCredential.fromJSON(object.data) : undefined,
    };
  },

  toJSON(message: AccountCredentialAuthResponse): unknown {
    const obj: any = {};
    if (message.status !== false) {
      obj.status = message.status;
    }
    if (message.code !== 0) {
      obj.code = Math.round(message.code);
    }
    if (message.msg !== "") {
      obj.msg = message.msg;
    }
    if (message.data !== undefined) {
      obj.data = IAccountCredential.toJSON(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialAuthResponse>, I>>(base?: I): AccountCredentialAuthResponse {
    return AccountCredentialAuthResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialAuthResponse>, I>>(
    object: I,
  ): AccountCredentialAuthResponse {
    const message = createBaseAccountCredentialAuthResponse();
    message.status = object.status ?? false;
    message.code = object.code ?? 0;
    message.msg = object.msg ?? "";
    message.data = (object.data !== undefined && object.data !== null)
      ? IAccountCredential.fromPartial(object.data)
      : undefined;
    return message;
  },
};

function createBaseAccountCredentialCreateResponse(): AccountCredentialCreateResponse {
  return { status: false, code: 0, msg: "", data: undefined };
}

export const AccountCredentialCreateResponse: MessageFns<AccountCredentialCreateResponse> = {
  encode(message: AccountCredentialCreateResponse, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.status !== false) {
      writer.uint32(8).bool(message.status);
    }
    if (message.code !== 0) {
      writer.uint32(16).int64(message.code);
    }
    if (message.msg !== "") {
      writer.uint32(26).string(message.msg);
    }
    if (message.data !== undefined) {
      IAccountCredential.encode(message.data, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AccountCredentialCreateResponse {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    let end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAccountCredentialCreateResponse();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.status = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.code = longToNumber(reader.int64());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.msg = reader.string();
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.data = IAccountCredential.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AccountCredentialCreateResponse {
    return {
      status: isSet(object.status) ? globalThis.Boolean(object.status) : false,
      code: isSet(object.code) ? globalThis.Number(object.code) : 0,
      msg: isSet(object.msg) ? globalThis.String(object.msg) : "",
      data: isSet(object.data) ? IAccountCredential.fromJSON(object.data) : undefined,
    };
  },

  toJSON(message: AccountCredentialCreateResponse): unknown {
    const obj: any = {};
    if (message.status !== false) {
      obj.status = message.status;
    }
    if (message.code !== 0) {
      obj.code = Math.round(message.code);
    }
    if (message.msg !== "") {
      obj.msg = message.msg;
    }
    if (message.data !== undefined) {
      obj.data = IAccountCredential.toJSON(message.data);
    }
    return obj;
  },

  create<I extends Exact<DeepPartial<AccountCredentialCreateResponse>, I>>(base?: I): AccountCredentialCreateResponse {
    return AccountCredentialCreateResponse.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<AccountCredentialCreateResponse>, I>>(
    object: I,
  ): AccountCredentialCreateResponse {
    const message = createBaseAccountCredentialCreateResponse();
    message.status = object.status ?? false;
    message.code = object.code ?? 0;
    message.msg = object.msg ?? "";
    message.data = (object.data !== undefined && object.data !== null)
      ? IAccountCredential.fromPartial(object.data)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}